Change the type used for time representation from double to long

+++ src/main/java/pixformer/controller/gameloop/TemplateGameLoop.java
@@ -6,14 +6,14 @@ package pixformer.controller.gameloop;
  */
 public abstract class TemplateGameLoop implements GameLoop {
 
-    private double previous;
+    private long previous;
 
     /**
      * When invoked, makes the game loop thread sleep for a certain amount of time.
      * 
      * @param dt the delta time between the previous iteration and the current one.
      */
-    abstract void sleep(double dt);
+    abstract void sleep(long dt);
 
     /**
      * When invoked the game loop processes the input of the user.
@@ -24,7 +24,7 @@ public abstract class TemplateGameLoop implements GameLoop {
      * Update the model.
      * @param dt dt the delta time between the previous iteration and the current one.
      */
-    abstract void update(double dt);
+    abstract void update(long dt);
 
     /**
      * Update the view.
@@ -33,7 +33,7 @@ public abstract class TemplateGameLoop implements GameLoop {
 
     @Override
     public final void loop(final long now) {
-        final double elapsed = now - previous;
+        final long elapsed = now - previous;
         processInput();
         update(elapsed);
         render();
